# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Fraud Detection & Alert System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /transactions:
    get:
      operationId: listTransactions
      tags: [transactions]
      summary: List transactions with optional filters
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: status
          in: query
          schema:
            type: string
            enum: [approved, blocked, flagged, pending_review]
        - name: riskLevel
          in: query
          schema:
            type: string
            enum: [low, medium, high, critical]
        - name: startDate
          in: query
          schema:
            type: string
            format: date-time
        - name: endDate
          in: query
          schema:
            type: string
            format: date-time
        - name: cardId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Paginated transaction list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedTransactions"

    post:
      operationId: scoreTransaction
      tags: [transactions]
      summary: Score and process a new transaction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScoreTransactionBody"
      responses:
        "200":
          description: Transaction scored
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionScoreResult"
  /transactions/{id}:
    get:
      operationId: getTransaction
      tags: [transactions]
      summary: Get a single transaction with full details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Transaction details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
  /transactions/{id}/review:
    post:
      operationId: reviewTransaction
      tags: [transactions]
      summary: Submit a manual review decision on a transaction
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewTransactionBody"
      responses:
        "200":
          description: Review submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"

  # ── Alerts ────────────────────────────────────────────────────────────────
...
  /rules:
    get:
      operationId: listRules
      tags: [rules]
      summary: List fraud detection rules
      responses:
        "200":
          description: Rules list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FraudRule"

    post:
      operationId: createRule
      tags: [rules]
      summary: Create a new fraud detection rule
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRuleBody"
      responses:
        "201":
          description: Rule created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FraudRule"
  /rules/{id}:
    patch:
      operationId: updateRule
      tags: [rules]
      summary: Update a fraud detection rule
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRuleBody"
      responses:
        "200":
          description: Rule updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FraudRule"

    delete:
      operationId: deleteRule
      tags: [rules]
      summary: Delete a fraud detection rule
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Rule deleted
  /rules/{id}/toggle:
    post:
      operationId: toggleRule
      tags: [rules]
      summary: Enable or disable a rule
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Rule toggled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FraudRule"

  # ── Users ─────────────────────────────────────────────────────────────────
...
  /merchants:
    get:
      operationId: listMerchants
      tags: [merchants]
      summary: List merchants
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: status
          in: query
          schema:
            type: string
            enum: [active, blocked, flagged, under_review]
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Merchant list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedMerchants"

    post:
      operationId: addMerchant
      tags: [merchants]
      summary: Add a merchant to the system
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddMerchantBody"
      responses:
        "201":
          description: Merchant added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Merchant"
  /merchants/{id}/block:
    post:
      operationId: blockMerchant
      tags: [merchants]
      summary: Block a merchant
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BlockMerchantBody"
      responses:
        "200":
          description: Merchant blocked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Merchant"
  /merchants/{id}/unblock:
    post:
      operationId: unblockMerchant
      tags: [merchants]
      summary: Unblock a merchant
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Merchant unblocked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Merchant"

  # ── SAR ───────────────────────────────────────────────────────────────────
  /sar:
    get:
      operationId: listSARs
      tags: [sar]
      summary: List Suspicious Activity Reports
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: status
          in: query
          schema:
            type: string
            enum: [draft, submitted, under_review, accepted, rejected]
      responses:
        "200":
          description: SAR list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedSARs"

    post:
      operationId: createSAR
      tags: [sar]
      summary: Create a new Suspicious Activity Report
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSARBody"
      responses:
        "201":
          description: SAR created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SAR"
  /sar/{id}:
    get:
      operationId: getSAR
      tags: [sar]
      summary: Get a SAR by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: SAR details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SAR"

    patch:
      operationId: updateSAR
      tags: [sar]
      summary: Update a SAR
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSARBody"
      responses:
        "200":
          description: SAR updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SAR"
  /sar/{id}/submit:
    post:
      operationId: submitSAR
      tags: [sar]
      summary: Submit SAR to regulators
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: SAR submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SAR"

  # ── Analytics / Dashboard ─────────────────────────────────────────────────
...
components:
  schemas:
...
    Transaction:
      type: object
      properties:
        id:
          type: string
        cardId:
          type: string
        userId:
          type: string
        merchantId:
          type: string
        merchantName:
          type: string
        amount:
          type: number
        currency:
          type: string
        status:
          type: string
          enum: [approved, blocked, flagged, pending_review]
        riskScore:
          type: number
          minimum: 0
          maximum: 100
        riskLevel:
          type: string
          enum: [low, medium, high, critical]
        fraudType:
          type: string
          nullable: true
        location:
          $ref: "#/components/schemas/Location"
        deviceFingerprint:
          type: string
          nullable: true
        responseTime:
          type: number
          description: Scoring latency in ms
        triggeredRules:
          type: array
          items:
            type: string
        anomalyScore:
          type: number
          nullable: true
        createdAt:
          type: string
          format: date-time
        reviewedBy:
          type: string
          nullable: true
        reviewedAt:
          type: string
          format: date-time
          nullable: true
        reviewDecision:
          type: string
          enum: [confirmed_fraud, false_positive, escalated]
          nullable: true
        reviewNotes:
          type: string
          nullable: true
      required: [id, cardId, userId, merchantId, merchantName, amount, currency, status, riskScore, riskLevel, createdAt]
    PaginatedTransactions:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Transaction"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
      required: [data, total, page, limit, totalPages]
    ScoreTransactionBody:
      type: object
      properties:
        cardId:
          type: string
        userId:
          type: string
        merchantId:
          type: string
        merchantName:
          type: string
        amount:
          type: number
        currency:
          type: string
        location:
          $ref: "#/components/schemas/Location"
        deviceFingerprint:
          type: string
          nullable: true
      required: [cardId, userId, merchantId, merchantName, amount, currency]
    TransactionScoreResult:
      type: object
      properties:
        transactionId:
          type: string
        decision:
          type: string
          enum: [approved, blocked, flagged, challenge_2fa]
        riskScore:
          type: number
        riskLevel:
          type: string
        responseTimeMs:
          type: number
        triggeredRules:
          type: array
          items:
            type: string
        explanation:
          type: string
      required: [transactionId, decision, riskScore, riskLevel, responseTimeMs, triggeredRules, explanation]
    ReviewTransactionBody:
      type: object
      properties:
        decision:
          type: string
          enum: [confirmed_fraud, false_positive, escalated]
        notes:
          type: string
        reviewerId:
          type: string
      required: [decision, reviewerId]
    Location:
      type: object
      properties:
        country:
          type: string
        city:
          type: string
        lat:
          type: number
        lng:
          type: number
        ipAddress:
          type: string
          nullable: true
      required: [country, city, lat, lng]

    # ── Alert ──────────────────────────────────────────────────────────────
...
    FraudRule:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        type:
          type: string
          enum: [velocity, geolocation, amount_threshold, device, pattern, blacklist]
        enabled:
          type: boolean
        priority:
          type: integer
        action:
          type: string
          enum: [block, flag, challenge_2fa, notify]
        conditions:
          type: object
          additionalProperties: true
        triggeredCount:
          type: integer
        falsePositiveRate:
          type: number
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, name, description, type, enabled, priority, action, conditions, triggeredCount, createdAt, updatedAt]
    CreateRuleBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        type:
          type: string
          enum: [velocity, geolocation, amount_threshold, device, pattern, blacklist]
        priority:
          type: integer
        action:
          type: string
          enum: [block, flag, challenge_2fa, notify]
        conditions:
          type: object
          additionalProperties: true
      required: [name, description, type, priority, action, conditions]
    UpdateRuleBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        enabled:
          type: boolean
        priority:
          type: integer
        action:
          type: string
          enum: [block, flag, challenge_2fa, notify]
        conditions:
          type: object
          additionalProperties: true

    # ── User ───────────────────────────────────────────────────────────────
...
    Merchant:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        category:
          type: string
        country:
          type: string
        status:
          type: string
          enum: [active, blocked, flagged, under_review]
        fraudRate:
          type: number
        totalTransactions:
          type: integer
        blockedReason:
          type: string
          nullable: true
        blockedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, name, category, country, status, fraudRate, totalTransactions, createdAt]
    PaginatedMerchants:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Merchant"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
      required: [data, total, page, limit, totalPages]
    AddMerchantBody:
      type: object
      properties:
        name:
          type: string
        category:
          type: string
        country:
          type: string
      required: [name, category, country]
    BlockMerchantBody:
      type: object
      properties:
        reason:
          type: string
      required: [reason]

    # ── SAR ────────────────────────────────────────────────────────────────
    SAR:
      type: object
      properties:
        id:
          type: string
        sarNumber:
          type: string
        caseId:
          type: string
        status:
          type: string
          enum: [draft, submitted, under_review, accepted, rejected]
        subjectName:
          type: string
        subjectType:
          type: string
          enum: [individual, business]
        reportingOfficer:
          type: string
        filingInstitution:
          type: string
        suspiciousActivityType:
          type: string
        description:
          type: string
        amountInvolved:
          type: number
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
        submittedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, sarNumber, caseId, status, subjectName, subjectType, reportingOfficer, filingInstitution, suspiciousActivityType, description, amountInvolved, startDate, endDate, createdAt, updatedAt]
    PaginatedSARs:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/SAR"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
      required: [data, total, page, limit, totalPages]
    CreateSARBody:
      type: object
      properties:
        caseId:
          type: string
        subjectName:
          type: string
        subjectType:
          type: string
          enum: [individual, business]
        reportingOfficer:
          type: string
        filingInstitution:
          type: string
        suspiciousActivityType:
          type: string
        description:
          type: string
        amountInvolved:
          type: number
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
      required: [caseId, subjectName, subjectType, reportingOfficer, filingInstitution, suspiciousActivityType, description, amountInvolved, startDate, endDate]
    UpdateSARBody:
      type: object
      properties:
        description:
          type: string
        amountInvolved:
          type: number
        suspiciousActivityType:
          type: string
        reportingOfficer:
          type: string

    # ── Dashboard / Analytics ───────────────────────────────────────────────
...